chore DOM logic
