+++ 11_To_Do/src/components/html/project-modal.html
@@ -7,7 +7,7 @@
         name="project-title"
         id="project-title"
         required
-        value="project 1"
+        value="Project 1"
       />
     </div>
     <div class="project-description">
