+++ 11_To_Do/src/index.html
@@ -8,10 +8,7 @@
   <body>
     <h1>To Do App</h1>
     <button class="create-project">Create New Project</button>
-    <div class="project">
-      <p>Project 1</p>
-      <button class="create-task" id="project1">Add New Task</button>
-    </div>
+    <div class="project"></div>
     <div id="taskDialogHTML"></div>
     <div id="projectDialogHTML"></div>
   </body>
